refactor all 'inline mode' (a.k.a. workflow mode) stuff into its own component `InlineMode`
