+++ client/src/app.tsx
@@ -13,17 +13,11 @@ import {
   useSubscription,
 } from "@apollo/client";
 import {
-  gqlGetItemCounts,
   gqlGetMyUser,
   gqlOnManuallyOpenedPinboardIdsChanged,
   gqlSetWebPushSubscriptionForUser,
 } from "../gql";
-import {
-  Item,
-  MyUser,
-  PinboardIdWithItemCounts,
-  User,
-} from "../../shared/graphql/graphql";
+import { Item, MyUser, User } from "../../shared/graphql/graphql";
 import { ItemWithParsedPayload } from "./types/ItemWithParsedPayload";
 import {
   desktopNotificationsPreferencesUrl,
@@ -46,10 +40,8 @@ import {
 } from "../../shared/constants";
 import { UserLookup } from "./types/UserLookup";
 import { gqlGetUsers } from "../gql";
-import {
-  InlinePinboardTogglePortal,
-  WORKFLOW_TITLE_QUERY_SELECTOR,
-} from "./inlinePinboardToggle";
+import { WORKFLOW_TITLE_QUERY_SELECTOR } from "./inlinePinboardToggle";
+import { InlineMode } from "./inlineMode";
 import { getAgateFontFaceIfApplicable } from "../fontNormaliser";
 import { Global } from "@emotion/react";
 
@@ -62,6 +54,11 @@ interface PinBoardAppProps {
 }
 
 export const PinBoardApp = ({ apolloClient, userEmail }: PinBoardAppProps) => {
+  const isInlineMode = useMemo(
+    () => window.location.hostname.startsWith("workflow."),
+    []
+  );
+
   const [payloadToBeSent, setPayloadToBeSent] = useState<PayloadAndType | null>(
     null
   );
@@ -71,33 +68,6 @@ export const PinBoardApp = ({ apolloClient, userEmail }: PinBoardAppProps) => {
   const [workflowTitleElements, setWorkflowTitleElements] = useState<
     HTMLElement[]
   >([]);
-
-  const workflowTitleElementLookup = useMemo(
-    () =>
-      workflowTitleElements.reduce((acc, node) => {
-        const pinboardId = node.parentElement?.id?.replace("stub-", "");
-        return pinboardId ? { ...acc, [pinboardId]: node } : acc;
-      }, {} as Record<string, HTMLElement>),
-    [workflowTitleElements]
-  );
-
-  // FIXME: move to a dedicated component
-  const itemCountsQuery = useQuery(gqlGetItemCounts, {
-    client: apolloClient,
-    variables: { pinboardIds: Object.keys(workflowTitleElementLookup) },
-    pollInterval: 5000,
-  });
-
-  const itemCounts: PinboardIdWithItemCounts[] =
-    itemCountsQuery.data?.getItemCounts || [];
-
-  const itemCountsLookup = itemCounts.reduce(
-    (lookup, element: PinboardIdWithItemCounts) => ({
-      ...lookup,
-      [element.pinboardId]: element,
-    }),
-    {} as Record<string, PinboardIdWithItemCounts>
-  );
 
   const queryParams = new URLSearchParams(window.location.search);
   // using state here but without setter, because host application/SPA might change url
@@ -405,18 +375,8 @@ export const PinBoardApp = ({ apolloClient, userEmail }: PinBoardAppProps) => {
             }
           >
             <TickContext.Provider value={lastTickTimestamp}>
-              {workflowTitleElements.length > 0 ? (
-                Object.entries(
-                  workflowTitleElementLookup
-                ).map(([pinboardId, node], index) => (
-                  <InlinePinboardTogglePortal
-                    key={index}
-                    node={node}
-                    pinboardId={pinboardId}
-                    counts={itemCountsLookup[pinboardId]}
-                    isLoading={itemCountsQuery.loading}
-                  />
-                ))
+              {isInlineMode ? (
+                <InlineMode workflowTitleElements={workflowTitleElements} />
               ) : (
                 <React.Fragment>
                   <Floaty isDropTarget={isDropTarget} />
